# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "ChatResponse": {
        "additionalProperties": false,
        "properties": {
          "delivery_error": {
            "anyOf": [
              {
                "enum": [
                  "pipefacil_api_key_missing",
                  "recipient_phone_missing",
                  "response_text_empty",
                  "response_media_url_missing",
                  "response_media_url_invalid",
                  "response_media_type_unsupported",
                  "pipefacil_transport_error",
                  "pipefacil_upstream_error"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Error"
          },
          "delivery_status": {
            "anyOf": [
              {
                "enum": [
                  "sent",
                  "failed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Status"
          },
          "intent": {
            "anyOf": [
              {
                "enum": [
                  "greeting",
                  "question",
                  "request",
                  "fallback"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          },
          "intent_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent Reason"
          },
          "response_messages": {
            "items": {
              "type": "string"
            },
            "title": "Response Messages",
            "type": "array"
          },
          "response_parts": {
            "items": {
              "$ref": "#/components/schemas/ResponsePart"
            },
            "title": "Response Parts",
            "type": "array"
          },
          "response_text": {
            "title": "Response Text",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "thread_id",
          "response_text",
          "status"
        ],
        "title": "ChatResponse",
        "type": "object"
      },
      "ConversationResumeRequest": {
        "additionalProperties": false,
        "properties": {
          "channel_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel Id"
          },
          "contact_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Id"
          },
          "context": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "deal_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deal Id"
          },
          "deal_seq": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deal Seq"
          },
          "history_limit": {
            "default": 100,
            "maximum": 500.0,
            "minimum": 1.0,
            "title": "History Limit",
            "type": "integer"
          },
          "profile_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Name"
          },
          "recipient_phone": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipient Phone"
          },
          "send_response": {
            "default": true,
            "title": "Send Response",
            "type": "boolean"
          },
          "sender_phone_number_id": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Phone Number Id"
          },
          "thread_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "thread_id"
        ],
        "title": "ConversationResumeRequest",
        "type": "object"
      },
      "ConversationResumeResponse": {
        "additionalProperties": false,
        "properties": {
          "delivery_error": {
            "anyOf": [
              {
                "enum": [
                  "pipefacil_api_key_missing",
                  "recipient_phone_missing",
                  "response_text_empty",
                  "response_media_url_missing",
                  "response_media_url_invalid",
                  "response_media_type_unsupported",
                  "pipefacil_transport_error",
                  "pipefacil_upstream_error"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Error"
          },
          "delivery_status": {
            "anyOf": [
              {
                "enum": [
                  "sent",
                  "failed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Status"
          },
          "history_message_count": {
            "title": "History Message Count",
            "type": "integer"
          },
          "history_source": {
            "default": "pipefacil",
            "title": "History Source",
            "type": "string"
          },
          "intent": {
            "anyOf": [
              {
                "enum": [
                  "greeting",
                  "question",
                  "request",
                  "fallback"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent"
          },
          "intent_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Intent Reason"
          },
          "response_messages": {
            "items": {
              "type": "string"
            },
            "title": "Response Messages",
            "type": "array"
          },
          "response_parts": {
            "items": {
              "$ref": "#/components/schemas/ResponsePart"
            },
            "title": "Response Parts",
            "type": "array"
          },
          "response_text": {
            "title": "Response Text",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "thread_id": {
            "title": "Thread Id",
            "type": "string"
          }
        },
        "required": [
          "thread_id",
          "response_text",
          "status",
          "history_message_count"
        ],
        "title": "ConversationResumeResponse",
        "type": "object"
      },
      "EventChannelPayload": {
        "properties": {
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "phoneNumber": {
            "title": "Phonenumber",
            "type": "string"
          },
          "phoneNumberId": {
            "title": "Phonenumberid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "phoneNumberId",
          "phoneNumber"
        ],
        "title": "EventChannelPayload",
        "type": "object"
      },
      "EventContactPayload": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "required": [
          "id"
        ],
        "title": "EventContactPayload",
        "type": "object"
      },
      "EventDealPayload": {
        "additionalProperties": true,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seq"
          },
          "stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EventDealStagePayload"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id"
        ],
        "title": "EventDealPayload",
        "type": "object"
      },
      "EventDealStagePayload": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name"
        ],
        "title": "EventDealStagePayload",
        "type": "object"
      },
      "EventMessagePayload": {
        "additionalProperties": true,
        "properties": {
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "externalId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Externalid"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "media": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "timestamp"
        ],
        "title": "EventMessagePayload",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "MessageReceivedEventData": {
        "properties": {
          "channel": {
            "$ref": "#/components/schemas/EventChannelPayload"
          },
          "contact": {
            "$ref": "#/components/schemas/EventContactPayload"
          },
          "deal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EventDealPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "$ref": "#/components/schemas/EventMessagePayload"
          }
        },
        "required": [
          "message",
          "channel",
          "contact"
        ],
        "title": "MessageReceivedEventData",
        "type": "object"
      },
      "MessageReceivedEventRequest": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MessageReceivedEventData"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          },
          "type": {
            "const": "message.received",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "timestamp",
          "data"
        ],
        "title": "MessageReceivedEventRequest",
        "type": "object"
      },
      "ResponsePart": {
        "additionalProperties": false,
        "properties": {
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "media_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Media Id"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "type": {
            "enum": [
              "text",
              "image",
              "video",
              "audio",
              "document"
            ],
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "ResponsePart",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SDR Agent Template",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/conversations/resume": {
      "post": {
        "operationId": "resume_conversation_conversations_resume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConversationResumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationResumeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Conversation",
        "tags": [
          "conversations"
        ]
      }
    },
    "/events/message-received": {
      "post": {
        "operationId": "message_received_events_message_received_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageReceivedEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Message Received"
      }
    },
    "/generated-audio/{filename}": {
      "get": {
        "operationId": "generated_audio_generated_audio__filename__get",
        "parameters": [
          {
            "in": "path",
            "name": "filename",
            "required": true,
            "schema": {
              "title": "Filename",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generated Audio"
      }
    },
    "/health": {
      "get": {
        "operationId": "healthcheck_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Healthcheck Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Healthcheck"
      }
    },
    "/ready": {
      "get": {
        "operationId": "readiness_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Readiness Ready Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Readiness"
      }
    }
  },
  "servers": [
    {
      "description": "Local server using the production configuration",
      "url": "http://localhost:8000"
    }
  ]
}